refactor: more store refactors
